Hotfix: -fix cross product for Vector4 -fix x,y coordinates for ray generation

+++ include/Geometry/Object3D.h
@@ -10,5 +10,5 @@ public:
     virtual f32 intersect(Ray ray, Hit hit, f32 tmin) = 0;
     virtual void id() = 0;
 
-    f32 m_color[3]{0.0, 0.0, 0.0};
+    u8 m_color[3]{0, 0, 0};
 };
